feat(component-overview): refactoring import statements
