+++ component-overview/webapp/routes/index.jsx
@@ -1,0 +1,2 @@
+export { default as Home } from './Home';
+export { default as ViewExample } from './ViewExample';
